Sikkerhet API-rotes
Lagt til ekstra sikkerhet på api routes som krever innlogging / employee role

+++ app/api/booking/[bookingId]/route.ts
@@ -1,4 +1,6 @@
 import prisma from '@/app/lib/prisma';
+import { authOptions } from '@/app/utils/authOptions';
+import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
 
 export async function DELETE(
@@ -11,7 +13,22 @@ export async function DELETE(
 ) {
     try {
         const { bookingId } = await params;
-
+        const session = await getServerSession(authOptions);
+        const user = await prisma.user.findUnique({
+            where: {
+                id: parseInt(session.user.id),
+            },
+        });
+        if (!user || !session || user.role !== 'EMPLOYEE') {
+            return NextResponse.json(
+                {
+                    success: false,
+                },
+                {
+                    status: 401,
+                },
+            );
+        }
         await prisma.booking.delete({
             where: {
                 id: parseInt(bookingId),
